chore: refactor deposit, write-off, and payout request seeders
