+++ test/unit/controller/payout-request-controller.ts
@@ -25,7 +25,6 @@ import PayoutRequestController from '../../../src/controller/payout-request-cont
 import User, { UserType } from '../../../src/entity/user/user';
 import PayoutRequest from '../../../src/entity/transactions/payout/payout-request';
 import Database from '../../../src/database/database';
-import { seedPayoutRequests } from '../../seed-legacy';
 import PayoutRequestRequest from '../../../src/controller/request/payout-request-request';
 import TokenHandler from '../../../src/authentication/token-handler';
 import RoleManager from '../../../src/rbac/role-manager';
@@ -41,7 +40,7 @@ import { truncateAllTables } from '../../setup';
 import generateBalance, { finishTestDB } from '../../helpers/test-helpers';
 import BalanceService from '../../../src/service/balance-service';
 import { inUserContext, UserFactory } from '../../helpers/user-factory';
-import { RbacSeeder, UserSeeder } from '../../seed';
+import { PayoutRequestSeeder, RbacSeeder, UserSeeder } from '../../seed';
 
 describe('PayoutRequestController', () => {
   let ctx: {
@@ -77,7 +76,7 @@ describe('PayoutRequestController', () => {
     await truncateAllTables(connection);
 
     const users = await new UserSeeder().seedUsers();
-    const { payoutRequests } = await seedPayoutRequests(users);
+    const { payoutRequests } = await new PayoutRequestSeeder().seedPayoutRequests(users);
 
     const adminUser = users.filter((u) => u.type === UserType.LOCAL_ADMIN)[0];
     const localUser = users.filter((u) => u.type === UserType.LOCAL_USER)[0];
