+++ test/unit/controller/stripe-webhook-controller.ts
@@ -32,8 +32,7 @@ import { extractRawBody } from '../../../src/helpers/raw-body';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import StripeDeposit from '../../../src/entity/stripe/stripe-deposit';
-import { seedStripeDeposits } from '../../seed-legacy';
-import { UserSeeder } from '../../seed';
+import { DepositSeeder, UserSeeder } from '../../seed';
 
 describe('StripeWebhookController', async (): Promise<void> => {
   let shouldSkip: boolean;
@@ -77,7 +76,7 @@ describe('StripeWebhookController', async (): Promise<void> => {
     const roleManager = await new RoleManager().initialize();
 
     const users = await new UserSeeder().seedUsers();
-    const { stripeDeposits } = await seedStripeDeposits(users);
+    const { stripeDeposits } = await new DepositSeeder().seedStripeDeposits(users);
 
     const controller = new StripeWebhookController({ specification, roleManager });
     app.use(json({
